# route: optimize virtual service route generation