This PR fixes the following
- Moves `BuildHTTPRoutesForVirtualService` outside of for loop. I could not find a reason why it has to be loop.
- Changes `separateVSHostsAndServices` method to process non wildcard hosts first so that we take advantage of `serviceRegistry` map for them