@ramaraochavali 2020-04-13
our Route Generation Benchmark improvements before and after **first**  change
```
name                               old time/op    new time/op    delta
RouteGeneration/empty-12             4.01ms ±16%    2.68ms ±47%     ~     (p=0.056 n=5+5)
RouteGeneration/telemetry-12         4.60ms ±16%    3.02ms ±41%  -34.35%  (p=0.016 n=5+5)
RouteGeneration/virtualservice-12    4.44ms ±24%    2.73ms ±42%     ~     (p=0.056 n=5+5)

name                               old alloc/op   new alloc/op   delta
RouteGeneration/empty-12              709kB ± 0%     709kB ± 0%     ~     (p=0.841 n=5+5)
RouteGeneration/telemetry-12          765kB ± 0%     765kB ± 0%     ~     (p=0.198 n=5+5)
RouteGeneration/virtualservice-12     722kB ± 0%     723kB ± 0%     ~     (p=0.548 n=5+5)

name                               old allocs/op  new allocs/op  delta
RouteGeneration/empty-12              7.72k ± 0%     7.72k ± 0%     ~     (all equal)
RouteGeneration/telemetry-12          8.53k ± 0%     8.53k ± 0%     ~     (all equal)
RouteGeneration/virtualservice-12     7.97k ± 0%     7.97k ± 0%     ~     (all equal)
```