@howardjohn 2020-04-13
> -63.80%
Wow, great work